Disjunction of browser.wait conditions
Using until.or to include more than one possible text to be present
in an element. We can use browser.wait with a disjunction, rather
than an expect call with a complex condition in it.

Also removing the base.scenario.ts that we added, and is not necessary
any more.